+++ pkgs/code_assets/lib/code_assets.dart
@@ -2,15 +2,24 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+/// @docImport 'package:hooks/hooks.dart';
 /// @docImport 'src/code_assets/code_asset.dart';
 /// @docImport 'src/code_assets/config.dart';
 
-/// Code asset support for hook authors.
+/// This package provides the API for code assets to be used with
+/// [`package:hooks`](https://pub.dev/packages/hooks).
 ///
-/// A code asset is an asset containing executable code which respects the
-/// native application binary interface (ABI).
+/// A [CodeAsset] is an asset containing executable code which respects the
+/// native application binary interface (ABI). These assets are bundled with a
+/// Dart or Flutter application. They can be produced by compiling C, C++,
+/// Objective-C, Rust, or Go code for example.
 ///
-/// Code assets can be added in a build hook as follows:
+/// This package is used in a build hook (`hook/build.dart`) to inform the Dart
+/// and Flutter SDKs about the code assets that need to be bundled with an
+/// application.
+///
+/// A [CodeAsset] can be added to the [BuildOutputBuilder] in a build hook as
+/// follows:
 ///
 /// <!-- file://./../example/api/code_assets_snippet.dart -->
 /// ```dart
@@ -34,13 +43,31 @@
 /// }
 /// ```
 ///
-/// See [CodeAsset] and [BuildOutputCodeAssetBuilder.add] for more details.
+/// The [CodeConfig] nested in the [HookInput] gives access to configuration
+/// specifically for compiling code assets. For example [CodeConfig.targetOS]
+/// and [CodeConfig.targetArchitecture] give access to the target OS and
+/// architecture that the code assets are compiled for:
 ///
-/// For more documentation of hooks, refer to the API docs of
-/// [`package:hooks`](https://pub.dev/packages/hooks).
+/// <!-- file://./../example/api/code_config_snippet.dart -->
+/// ```dart
+/// import 'package:code_assets/code_assets.dart';
+/// import 'package:hooks/hooks.dart';
 ///
-/// When compiling C, C++ or Objective-C code from source, consider using
-/// [`package:native_toolchain_c`](https://pub.dev/packages/native_toolchain_c).
+/// void main(List<String> args) async {
+///   await build(args, (input, output) async {
+///     if (input.config.buildCodeAssets) {
+///       final codeConfig = input.config.code;
+///       final targetOS = codeConfig.targetOS;
+///       final targetArchitecture = codeConfig.targetArchitecture;
+///
+///       // Add some code assets.
+///     }
+///   });
+/// }
+/// ```
+///
+/// For more information about build hooks see
+/// [dart.dev/tools/hooks](https://dart.dev/tools/hooks).
 library;
 
 export 'src/code_assets/architecture.dart' show Architecture;