+++ pkgs/hooks/lib/src/config.dart
@@ -7,6 +7,7 @@ import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:crypto/crypto.dart' show sha256;
+import 'package:meta/meta.dart';
 
 import 'api/build_and_link.dart';
 import 'encoded_asset.dart';
@@ -28,7 +29,15 @@ sealed class HookInput {
   /// The directory in which output and intermediate artifacts that are unique
   /// to the [config] can be placed.
   ///
-  /// This directory is unique per hook and per [config].
+  /// This directory is unique per hook and per [config]. The directory is
+  /// nested inside [outputDirectoryShared] and has a short checksum to avoid
+  /// running out of path length on Winddows.
+  ///
+  /// Prefer using a sub directory in [outputDirectoryShared] with a
+  /// checksum of the fields on the [config] that influence your build. Reusing
+  /// a precise subdirectory only dependent on what influences your build avoids
+  /// cache misses for [config]s that differ in fields irrelevant for your
+  /// build.
   ///
   /// The contents of this directory will not be modified by anything else than
   /// the hook itself.
@@ -55,7 +64,12 @@ sealed class HookInput {
   /// The directory in which shared output and intermediate artifacts can be
   /// placed.
   ///
-  /// This directory is unique per hook.
+  /// This directory is unique per hook. Use a sub directory of
+  /// [outputDirectoryShared] with a checksum of the parts on the [config] that
+  /// influence your assets. Ensure your checksum is relatively short to avoid
+  /// running out of path lengths on Windows. Reusing a precise subdirectory
+  /// only dependent on what influences your build avoids cache misses for
+  /// [config]s that differ in fields irrelevant for your build.
   ///
   /// The contents of this directory will not be modified by anything else than
   /// the hook itself.
@@ -194,10 +208,12 @@ String _jsonChecksum(Map<String, Object?> json) {
   final hash = sha256
       .convert(const JsonEncoder().fuse(const Utf8Encoder()).convert(json))
       .toString()
-      // 256 bit hashes lead to 64 hex character strings.
-      // To avoid overflowing file paths limits, only use 32.
-      // Using 16 hex characters would also be unlikely to have collisions.
-      .substring(0, 32);
+      // 256 bit hashes lead to 64 hex character strings. To avoid overflowing
+      // file paths limits on Windows, only use 10. 10 hex characters with 1000
+      // different configs leads to a one in a million collision chance. On
+      // collision, we'd get a cache miss and rerun the hooks and throw away the
+      // cache for the colliding configuration.
+      .substring(0, 10);
   return hash;
 }
 
@@ -331,6 +347,12 @@ final class LinkInput extends HookInput {
   List<EncodedAsset> get _encodedAssets => assets.encodedAssets;
 
   /// The file containing recorded usages, if any.
+  ///
+  /// Experimental: The record uses feature needs to be enabled as experiment.
+  /// The experiment is only available in the Dart SDK, not in Flutter. We
+  /// reserve the right to break this API at any point without respecting
+  /// semantic versioning of this package.
+  @experimental
   Uri? get recordedUsagesFile => _syntaxLinkInput.resourceIdentifiers;
 
   @override